PROF-54: base CmsPersistentObjectBase class is extracted
